requirements.txt
flask==3.1.0
bcrypt==4.2.1
Pillow>=12.1.1,<13
PyJWT==2.8.0
python-json-logger
sentry-sdk
SQLAlchemy>=2.0.49,<2.1
alembic==1.11.1
redis>=4.6.0
fakeredis>=2.21.0
prometheus_client>=0.16.0
